Deterministic psuedo-random alloc for HBM Cache mode
When running in HBM cache mode it is desirable to not allocate
sequential physical pages due to the structure of the HBM cache.
The original mOS code called a randomizer to select physical pages,
however the run-to-run variation was high. This fix will use a
deterministic psuedo-random generator, in which the starting seed is the
same at the start of each process, whether that be processes in the same
job or processes in subsequent jobs. This will improve the variation
observed across processes and will satisfy the original intent of not
mapping an excessive number of pages to a single HBM cache line.

Signed-off-by: John Attinella <john.e.attinella@intel.com>