mOS: OneCCL worker affinity mask management
If left to its own defaults, OneCCL may attempt to affinitize worker
threads onto CPUs outside of the LWK CPUs that have been reserved for
the current process. This will result in an error being reported by the
runtime and the application will be terminated. The same problem will
occur if the user supplied an affinity list containing CPUs outside of
the reserved LWK CPUs. This change will correct these situations. If no
CCL_WORKER_AFFINITY list is provided, YOD will create one prior to
executing the application. It will assign the CCL worker threads to the
last 4 CPUs in the sequence list of reserved LWK CPUs. If a valid
affinity specification is provided by the caller, no change will be
performed. The environment variable that controls this is:
CCL_WORKER_AFFINITY

Signed-off-by: John Attinella <john.e.attinella@intel.com>